Move revisions folder to /lib
