+++ lib/revisions/routes.js
@@ -2,10 +2,10 @@ const express = require('express')
 const {isString, isPlainObject} = require('lodash')
 const createError = require('http-errors')
 const contentDisposition = require('content-disposition')
-const errorHandler = require('../lib/util/error-handler')
-const {isCommuneActuelle} = require('../lib/util/cog')
-const w = require('../lib/util/w')
-const rawBodyParser = require('../lib/util/raw-body-parser')
+const errorHandler = require('../util/error-handler')
+const {isCommuneActuelle} = require('../util/cog')
+const w = require('../util/w')
+const rawBodyParser = require('../util/raw-body-parser')
 const {validateBAL} = require('./validate-bal')
 const {fetchRevision, getCurrentRevision, getRevisionsByCommune, createRevision, setFile, getFiles, getFileData, publishRevision, computeRevision} = require('./model')
