+++ lib/revisions/validate-bal.js
